Improve structure of buildings
Include the building's name in the object, instead of it being a
property in a map.
When the client is playing a match, just return a list of the player's
team's buildings.
When the client is spectating a match, return a map of team to a list
of their buildings.